+++ src/routes/(v2)/v2/files-proxy/schema.ts
@@ -4,7 +4,8 @@ export const schema = z.object({
 	images: z
 		.custom<File>()
 		.refine((f) => f instanceof File && f.size < 10000, 'Max 10Kb upload size.')
-		.array(),
+		.array()
+		.nullable(),
 	image: z
 		.instanceof(File, { message: 'Please upload a file.' })
 		.refine((f) => f.size < 10000, 'Max 10Kb upload size.')